pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jane-street-prediction"
version = "0.1.0"
description = "Jane Street Market Prediction with Autoencoder-XGBoost Ensemble"
authors = [{name = "James Lee", email = "jamesle2@andrew.cmu.edu"}]
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.21.0", 
    "scikit-learn>=1.1.0",
    "xgboost>=1.6.0",
    "tensorflow>=2.9.0",
    "pydantic>=1.10.0",
    "pyyaml>=6.0",
    "jupyter>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0"
]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88